Feat/visit auth

+++ event-service/src/main/java/com/example/eventservice/repository/event/CulturalEventRepository.java
@@ -1,17 +1,14 @@
 package com.example.eventservice.repository.event;
 
 import com.example.eventservice.entity.event.CulturalEvent;
-import com.example.eventservice.entity.event.CulturalEventDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
 
 
-public interface CulturalEventRepository extends JpaRepository<CulturalEvent, Long>{
-
+public interface CulturalEventRepository extends JpaRepository<CulturalEvent, Integer>{
 
     @Query(value = "select * from cultural_event as c where c.id = :culturalEventId for update", nativeQuery = true)
     Optional<CulturalEvent> findByIdForUpdate(int culturalEventId);
